pyproject.toml
[tool.poetry]
name = "vcon-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
gunicorn = ">=23.0.0"
datadog = "^0.48.0"
uvicorn = "0.23.2"
redis = "4.6.0"
fastapi = "0.115.6"
python-json-logger = "2.0.4"
pyyaml = "^6.0.1"
peewee = ">=3.17.1"
sentry-sdk = ">=2.8.0"
pydantic = "^2.6.2"
tenacity = "^8.2.3"
uuid6 = "^2024.1.12"
pydash = "^7.0.7"
requests = "^2.31.0"
slack-sdk = "^3.27.1"
boto3 = "^1.34.52"
deepgram-sdk = "^3.1.5"
openai = ">=1.54.3"
groq = "^0.4.0"
psycopg2-binary = "^2.9.9"
pymongo = "^4.6.2"
elasticsearch = "^8.13.1"
pycose= "^1.0.1"
ecdsa= "^0.18.0"
python-dotenv = "^1.0.1"
starlette = ">=0.40.0"
transformers = "^4.48.0"
torch = "^2.2.1"
trio = "^0.28.0"
pytest = "^8.3.4"
anyio = "^4.8.0"
faker = "^33.3.1"
jq = "^1.8.0"
pytest-redis = "^3.1.3"
pymilvus = ">=2.3.0"
python-dateutil = "^2.8.2"
pytest-dotenv = "^0.5.2"
msal = "^1.32.3"
azure-storage-blob = "^12.25.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.0.2"
pytest-asyncio = "^0.23.5"
pytest-dotenv = "^0.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 120
skip-string-normalization = true